proxy/requirements.txt
fastapi
uvicorn
httpx==0.28.1
pydantic
slowapi
pdf2image
pillow
python-multipart